# Conflicts:
#	modules/core/src/main/java/org/apache/ignite/internal/processors/platform/client/ClientResponse.java

+++ modules/core/src/main/java/org/apache/ignite/internal/processors/platform/client/ClientBooleanResponse.java
@@ -31,7 +31,7 @@ public class ClientBooleanResponse extends ClientResponse {
      *
      * @param requestId Request id.
      */
-    public ClientBooleanResponse(int requestId, boolean val) {
+    public ClientBooleanResponse(long requestId, boolean val) {
         super(requestId);
 
         this.val = val;
